clear implicit conversion warnings
reported by clang-11 -fsanitize=integer

mkvparser/mkvparser.cc:301:31: runtime error: implicit conversion from
type 'signed char' of value -128 (8-bit, signed) to type 'unsigned long
long' changed the value to 18446744073709551488 (64-bit, unsigned)

m2ts/webm2pes.cc:71:13: runtime error: implicit conversion from type
'unsigned int' of value 1260 (32-bit, unsigned) to type 'std::uint8_t'
(aka 'unsigned char') changed the value to 236 (8-bit, unsigned)

+++ mkvparser/mkvparser.cc
@@ -298,7 +298,7 @@ long UnserializeInt(IMkvReader* pReader, long long pos, long long size,
   if (status < 0)
     return status;
 
-  unsigned long long result = first_byte;
+  unsigned long long result = static_cast<unsigned long long>(first_byte);
   ++pos;
 
   for (long i = 1; i < size; ++i) {
